NT208/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.2
openai==1.51.0
python-dotenv==1.0.1
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.5.2
bcrypt==4.2.0

